Fix KBinsDiscretizer xfails

+++ python/cuml/cuml/_thirdparty/sklearn/preprocessing/_discretization.py
@@ -250,7 +250,7 @@ class KBinsDiscretizer(TransformerMixin, BaseEstimator, SparseInputTagMixin):
             )
             # Fit the OneHotEncoder with toy datasets
             # so that it's ready for use after the KBinsDiscretizer is fitted
-            self._encoder.fit(np.zeros((1, len(self.n_bins_)), dtype=int))
+            self._encoder.fit(np.zeros((1, len(self.n_bins_))))
 
         return self
